add separate modules for color and letterSpacing generators
