+++ packages/crayon/src/plugin/generators/__tests__/color.test.ts
@@ -1,0 +1,19 @@
+import { color } from '../color'
+
+jest.mock('../../../utils/tailwindUtils')
+
+test('generates color property with tailwind values', () => {
+  expect(color('blue-500'))
+    .toEqual({ color: '#3b82f6' })
+
+  expect(color('transparent'))
+    .toEqual({ color: 'transparent' })
+
+  expect(color('current'))
+    .toEqual({ color: 'currentColor' })
+})
+
+test('uses value as raw css if it cant be resolved from tailwind', () => {
+  expect(color('#10b981'))
+    .toEqual({ color: '#10b981' })
+})
