+++ packages/crayon/src/plugin/generators/letterSpacing.ts
@@ -1,0 +1,8 @@
+import { Generator } from '../../types'
+import resolveThemeValue from '../resolveThemeValue'
+
+export const letterSpacing: Generator = (value) => {
+  return {
+    letterSpacing: resolveThemeValue(`letterSpacing.${value}`, value),
+  }
+}
